requirements.txt
# requirements.txt
boto3==1.35.0
pandas==2.2.0
pyarrow==15.0.0
faker==24.0.0
python-dotenv==1.0.0
boto3-stubs[s3,kinesis,glue,athena,cloudwatch]==1.35.0
